+++ src/types.ts
@@ -1,17 +1,13 @@
 import type {
-  EmbeddingModelV3 as AIEmbeddingModelV3,
+  EmbeddingModelV3,
   LanguageModelV3,
   LanguageModelV3CallOptions,
   LanguageModelV3StreamPart,
 } from '@ai-sdk/provider';
 import type { ProviderOptions } from '@ai-sdk/provider-utils';
 
-export type EmbeddingModelV3<VALUE = any> = AIEmbeddingModelV3<VALUE>;
-export type { LanguageModelV3 };
-
-export type EmbeddingModel<VALUE = any> = AIEmbeddingModelV3<VALUE>;
 export type LanguageModel = LanguageModelV3;
-
+export type EmbeddingModel<VALUE = unknown> = EmbeddingModelV3<VALUE>;
 export type LanguageModelCallOptions = LanguageModelV3CallOptions;
 export type LanguageModelStreamPart = LanguageModelV3StreamPart;
 
@@ -19,7 +15,7 @@ export type LanguageModelStreamPart = LanguageModelV3StreamPart;
  * Options for creating a retryable model.
  */
 export interface RetryableModelOptions<
-  MODEL extends LanguageModel | EmbeddingModel<any>,
+  MODEL extends LanguageModel | EmbeddingModel,
 > {
   model: MODEL;
   retries: Retries<MODEL>;
@@ -31,7 +27,7 @@ export interface RetryableModelOptions<
 /**
  * The context provided to Retryables with the current attempt and all previous attempts.
  */
-export type RetryContext<MODEL extends LanguageModel | EmbeddingModel<any>> = {
+export type RetryContext<MODEL extends LanguageModel | EmbeddingModel> = {
   /**
    * Current attempt that caused the retry
    */
@@ -45,9 +41,7 @@ export type RetryContext<MODEL extends LanguageModel | EmbeddingModel<any>> = {
 /**
  * A retry attempt with an error
  */
-export type RetryErrorAttempt<
-  MODEL extends LanguageModel | EmbeddingModel<any>,
-> = {
+export type RetryErrorAttempt<MODEL extends LanguageModel | EmbeddingModel> = {
   type: 'error';
   error: unknown;
   result?: undefined;
@@ -67,14 +61,14 @@ export type RetryResultAttempt = {
 /**
  * A retry attempt with either an error or a result and the model used
  */
-export type RetryAttempt<MODEL extends LanguageModel | EmbeddingModel<any>> =
+export type RetryAttempt<MODEL extends LanguageModel | EmbeddingModel> =
   | RetryErrorAttempt<MODEL>
   | RetryResultAttempt;
 
 /**
  * A model to retry with and the maximum number of attempts for that model.
  */
-export type Retry<MODEL extends LanguageModel | EmbeddingModel<any>> = {
+export type Retry<MODEL extends LanguageModel | EmbeddingModel> = {
   model: MODEL;
   maxAttempts?: number;
   delay?: number;
@@ -86,17 +80,16 @@ export type Retry<MODEL extends LanguageModel | EmbeddingModel<any>> = {
 /**
  * A function that determines whether to retry with a different model based on the current attempt and all previous attempts.
  */
-export type Retryable<MODEL extends LanguageModel | EmbeddingModel<any>> = (
+export type Retryable<MODEL extends LanguageModel | EmbeddingModel> = (
   context: RetryContext<MODEL>,
-) => Retry<MODEL> | Promise<Retry<MODEL>> | undefined;
+) => Retry<MODEL> | Promise<Retry<MODEL> | undefined> | undefined;
 
-export type Retries<MODEL extends LanguageModel | EmbeddingModel<any>> = Array<
+export type Retries<MODEL extends LanguageModel | EmbeddingModel> = Array<
   Retryable<MODEL> | Retry<MODEL> | MODEL
 >;
 
-export type RetryableOptions<
-  MODEL extends LanguageModel | EmbeddingModel<any>,
-> = Partial<Omit<Retry<MODEL>, 'model'>>;
+export type RetryableOptions<MODEL extends LanguageModel | EmbeddingModel> =
+  Partial<Omit<Retry<MODEL>, 'model'>>;
 
 export type LanguageModelGenerate = Awaited<
   ReturnType<LanguageModel['doGenerate']>
@@ -110,6 +103,6 @@ export type EmbeddingModelCallOptions<VALUE> = Parameters<
   EmbeddingModel<VALUE>['doEmbed']
 >[0];
 
-export type EmbeddingModelEmbed<VALUE> = Awaited<
+export type EmbeddingModelEmbed<VALUE = any> = Awaited<
   ReturnType<EmbeddingModel<VALUE>['doEmbed']>
 >;